requirements.txt
fastapi===0.115.11
uvicorn
gunicorn
timescaledb
sqlmodel
pydantic
sqlalchemy
ipykernel
requests
httpx
python-decouple
psycopg[binary]
faker